check for Ubuntu account before removing it!

+++ container-creation/image-installers/r-ai-d-base.sh
@@ -6,8 +6,11 @@ source set_container_envars
 
 echo "** R AI Distrobox Base **"
 
-echo "..Removing 'ubuntu' user" 1>&2
-userdel --remove ubuntu 2> /dev/null
+if [[ "$(grep ubuntu /etc/passwd)" =~ "1000" ]]
+then
+  echo "..Removing 'ubuntu' user" 1>&2
+  userdel --remove ubuntu 2> /dev/null
+fi
 
 echo "..Restoring documentation" 1>&2
 export DEBIAN_FRONTEND=noninteractive
